Shuffle types around so Interfaces.ts imports less code
