+++ src/models/TransactionType.ts
@@ -1,0 +1,5 @@
+export enum TransactionType {
+  Swap = 'swap',
+  Deposit = 'deposit',
+  Remove = 'remove',
+}
