+++ src/utils/postgresTables/mevDetection/Sandwich/SandwichUtils.ts
@@ -1,16 +1,18 @@
-import { Op } from "sequelize";
-import { TransactionCoins } from "../../../../models/TransactionCoins.js";
-import { TransactionData } from "../../../../models/Transactions.js";
-import { ExtendedTransactionData, SandwichLoss, TransactionCoin, TransactionCoinRecord } from "../../../Interfaces.js";
-import { findCoinSymbolById } from "../../readFunctions/Coins.js";
-import { WEB3_HTTP_PROVIDER, getTokenTransferEvents, getTxFromTxId } from "../../../web3Calls/generic.js";
-import { getAbiBy } from "../../Abi.js";
-import { LossTransaction, Sandwiches } from "../../../../models/Sandwiches.js";
-import { readSandwichesInBatches, readSandwichesInBatchesForBlock } from "../../readFunctions/Sandwiches.js";
-import { calculateLossForDeposit, calculateLossForSwap, calculateLossForWithdraw } from "./VictimLossFromSandwich.js";
-import { IsSandwich } from "../../../../models/IsSandwich.js";
+import { Op } from 'sequelize';
+import { TransactionCoins } from '../../../../models/TransactionCoins.js';
+import { ExtendedTransactionData, TransactionCoin, TransactionData } from '../../../../models/Transactions.js';
+import { SandwichLoss, TransactionCoinRecord } from '../../../Interfaces.js';
+import { findCoinSymbolById } from '../../readFunctions/Coins.js';
+import { WEB3_HTTP_PROVIDER, getTokenTransferEvents, getTxFromTxId } from '../../../web3Calls/generic.js';
+import { getAbiBy } from '../../Abi.js';
+import { LossTransaction, Sandwiches } from '../../../../models/Sandwiches.js';
+import { readSandwichesInBatches, readSandwichesInBatchesForBlock } from '../../readFunctions/Sandwiches.js';
+import { calculateLossForDeposit, calculateLossForSwap, calculateLossForWithdraw } from './VictimLossFromSandwich.js';
+import { IsSandwich } from '../../../../models/IsSandwich.js';
 
-export async function enrichCandidateWithCoinInfo(candidate: TransactionData[]): Promise<ExtendedTransactionData[] | null> {
+export async function enrichCandidateWithCoinInfo(
+  candidate: TransactionData[]
+): Promise<ExtendedTransactionData[] | null> {
   // Extract tx_ids from candidate array
   const txIds = candidate.map((transaction) => transaction.tx_id);
