+++ src/utils/postgresTables/txParsing/ParseRemoveLiquidity.ts
@@ -1,5 +1,5 @@
 import { saveCoins, saveTransaction } from './ParsingHelper.js';
-import { TransactionType } from '../../../models/Transactions.js';
+import { TransactionType } from '../../../models/TransactionType.js';
 import { getCoinsBy } from '../readFunctions/Pools.js';
 import { getCoinIdByAddress, findCoinDecimalsById } from '../readFunctions/Coins.js';
 import { Coin } from '../../Interfaces.js';
