+++ src/utils/postgresTables/txParsing/ParseTokenExchangeUnderlying.ts
@@ -1,5 +1,5 @@
 import { saveCoins, saveTransaction } from './ParsingHelper.js';
-import { TransactionType } from '../../../models/Transactions.js';
+import { TransactionType } from '../../../models/TransactionType.js';
 import { getCoinsBy, getPoolIdByPoolAddress, getBasePoolBy } from '../readFunctions/Pools.js';
 import { getCoinIdByAddress, findCoinDecimalsById } from '../readFunctions/Coins.js';
 import { findTransactionsByPoolIdAndHash } from '../readFunctions/Transactions.js';
